split gulpfile to tasks
