+++ gulp/tasks/buildTest.js
@@ -1,0 +1,11 @@
+var gulp = require('gulp');
+
+var buildEnd = require('../util/buildEnd.js');
+
+gulp.task('buildTest', [
+    'buildClean',
+    'buildScripts',
+    'buildTestStyles',
+    'doc',
+    'copyPrivateAssets'
+], buildEnd);
