+++ gulp/tasks/buildTestStyles.js
@@ -1,0 +1,9 @@
+var gulp = require('gulp');
+
+var destCSS = require('../util/destCSS');
+
+gulp.task('buildTestStyles', ['buildClean'], function (cb) {
+    destCSS({
+        isTest: true
+    }, cb);
+});
