+++ gulp/tasks/collectImagesUsageStats.js
@@ -1,0 +1,53 @@
+var header = require('gulp-header');
+var rename = require('gulp-rename');
+var es = require('event-stream');
+var less = require('gulp-less');
+var gulp = require('gulp');
+var path = require('path');
+var glob = require('glob');
+
+var error = require('../util/error');
+var config = require('../../build/config');
+var deps = require('../../build/gulp-deps')(config);
+
+gulp.task('collectImagesUsageStats', ['buildClean', 'cleanupTmpLess'], function (cb) {
+    var skins = deps.getSkinsList();
+
+    var imagesBasePath = path.resolve(__dirname + '/build/tmp/img_all');
+
+    var statisticsStreams = skins.map(function (skinName) {
+        var skinLessFiles = glob.sync('./src/**/' + skinName + '/less/*.less');
+
+        skinLessFiles.unshift('./private/less/mixins.images-usage-statistics.less');
+        skinLessFiles.unshift('./private/less/mixins.ie8.less');
+
+        skinLessFiles = skinLessFiles.map(function (lessFilePath) {
+            return lessFilePath + ':reference';
+        });
+
+        return gulp.src('private/less/images-usage-statistics.less')
+            .pipe(error.handle())
+            .pipe(header(deps.lessHeader({
+                variables: {
+                    baseURL: '\'__BASE_URL__\'',
+                    spritesURL: '\'__BASE_URL__\'',
+
+                    ie8: true,
+                    useSprites: true,
+                    mobile: false,
+                    retina: false,
+
+                    skinName: skinName,
+
+                    imagesBasePath: '\'' + imagesBasePath + '\''
+                },
+                imports: skinLessFiles
+            })))
+            .pipe(less())
+            .pipe(rename('images-usage-statistics.' + skinName + '.less'))
+            .pipe(gulp.dest('build/tmp/less/'));
+    });
+
+    var stream = es.concat.apply(null, statisticsStreams);
+    stream.on('end', cb);
+});
