+++ gulp/tasks/copySprites.js
@@ -1,0 +1,9 @@
+var gulp = require('gulp');
+
+var error = require('../util/error');
+
+gulp.task('copySprites', ['buildClean', 'copySVG', 'generateSprites'], function () {
+    return gulp.src('build/tmp/img/sprite*.png')
+        .pipe(error.handle())
+        .pipe(gulp.dest('public/img'));
+});
