+++ gulp/tasks/doc.js
@@ -1,0 +1,10 @@
+var gulp = require('gulp');
+
+var gendoc = require('../../docbuilder/gendoc');
+var config = require('../../build/config');
+
+gulp.task('doc', function () {
+    var doc = config.doc;
+
+    gendoc.generateDocumentation(doc.menu, doc.input, doc.output);
+});
