+++ gulp/tasks/watch.js
@@ -1,0 +1,7 @@
+var gulp = require('gulp');
+
+gulp.task('watch', function () {
+    gulp.watch('private/*.*', ['copyPrivateAssets']);
+    gulp.watch('vendors/leaflet/src/**/*.*', ['buildLeaflet']);
+    gulp.watch('src/doc/**/*.*', ['doc']);
+});
