+++ gulp/util/projectList.js
@@ -1,0 +1,10 @@
+var projectList;
+
+module.exports = {
+    set: function(val) {
+        projectList = val;
+    },
+    get: function() {
+        return projectList;
+    }
+};
