+++ package.json
@@ -39,7 +39,6 @@
     "gulp-jshint": "1.6.4",
     "gulp-karma": "0.0.4",
     "gulp-less": "1.3.3",
-    "gulp-load-plugins": "0.5.0",
     "gulp-minify-css": "0.3.2",
     "gulp-newer": "^0.4.0",
     "gulp-notify": "1.4.0",
@@ -64,7 +63,7 @@
     "optimist": "^0.6.1",
     "pretty-bytes": "0.1.0",
     "request": "2.33.0",
-    "run-sequence": "0.3.6",
+    "require-dir": "^0.1.0",
     "through2": "0.4.1",
     "underscore.string": "2.3.3"
   },
